% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{add_step}
\alias{add_step}
\title{Add a step to a workflow}
\usage{
add_step(x, new_step, position, ...)
}
\arguments{
\item{x}{The workflow object to which the step will be added.}

\item{new_step}{The step to add to the workflow.}

\item{position}{The position in the workflow where the new step should be added.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The updated workflow object with the new step added.
}
\description{
Generic method to add a step to a workflow object.
}
