% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-mutate.R
\name{add_step.workflow}
\alias{add_step.workflow}
\title{Add a new step to the workflow}
\usage{
\method{add_step}{workflow}(x, new_step, position = length(x$steps) + 1L, ...)
}
\arguments{
\item{x}{The `workflow` object to update.}

\item{new_step}{A `workflowstep` object to add to the workflow.}

\item{position}{An integer index specifying where to insert the new step (default: at the end
of the steps list).}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflow` object with the new step added.
}
\description{
This method adds a new `workflowstep` to the `steps` list of a `workflow` object at a specified
position. It also performs validation after adding the step and updates the current index and
entries of all steps to maintain consistency.
}
