% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{as.commands_record}
\alias{as.commands_record}
\title{Convert an object to commands.json record format}
\usage{
as.commands_record(x, ...)
}
\arguments{
\item{x}{The object to convert.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic method to create records matching the commands.json schema.
}
