% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{as.commands_record.workflow}
\alias{as.commands_record.workflow}
\title{Convert a workflow object to commands.json record format}
\usage{
\method{as.commands_record}{workflow}(x, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A list summarizing workflow steps for commands.json.
}
\description{
This method converts a `workflow` object into a list format suitable for writing to a
`commands.json` file, which is used to define the workflow steps in a structured way.
It applies the conversion to each step in the workflow.
}
