% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{as.data.frame.workflow}
\alias{as.data.frame.workflow}
\title{Convert a workflow object to a data frame}
\usage{
\method{as.data.frame}{workflow}(x, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame summarizing the workflow steps.
}
\description{
This method converts a `workflow` object into a data frame summarizing its steps.
}
