% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowrun-class.R
\name{as.data.frame.workflowrun}
\alias{as.data.frame.workflowrun}
\title{Convert a workflowrun object to a data frame}
\usage{
\method{as.data.frame}{workflowrun}(x, ...)
}
\arguments{
\item{x}{A `workflowrun` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame summarizing the workflow run state.
}
\description{
This method converts a `workflowrun` object into a data frame summarizing its state.
}
