% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstep-class.R
\name{as.data.frame.workflowstep}
\alias{as.data.frame.workflowstep}
\title{Convert a workflowstep to a data frame}
\usage{
\method{as.data.frame}{workflowstep}(x, ...)
}
\arguments{
\item{x}{A `workflowstep` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame with one row representing the workflow step.
}
\description{
This method converts a `workflowstep` object into a data frame format, which can be
useful for tabular display or exporting.
}
