% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-config.R
\name{config}
\alias{config}
\title{Load PEITHO configuration}
\usage{
config()
}
\value{
A list containing the configuration values.
}
\description{
Loads the PEITHO configuration from the `config.yaml` file in the package.
}
