% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{current_step.workflow}
\alias{current_step.workflow}
\title{Accessor functions ------------------------------------------------------
Get the current step of the workflow}
\usage{
\method{current_step}{workflow}(x, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{...}{Additional arguments (not used).}
}
\value{
The current `workflow_step` object, or `NULL` if no current step.
}
\description{
Accessor functions ------------------------------------------------------
Get the current step of the workflow
}
