% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-operationparam-class.R
\name{extract_arg.operationparam}
\alias{extract_arg.operationparam}
\title{Extract argument value from operationparam object}
\usage{
\method{extract_arg}{operationparam}(x, last_result = list(""), ...)
}
\arguments{
\item{x}{A `operationparam` object.}

\item{last_result}{The result from the previous step, used if the parameter type is "result".}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing the argument value, named if applicable.
}
\description{
Extract argument value from operationparam object
}
