% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{extract_inputs}
\alias{extract_inputs}
\title{Extract inputs from a workflow}
\usage{
extract_inputs(object, ...)
}
\arguments{
\item{object}{A `workflow` object or a `workflowrun` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A named list of input values.
}
\description{
This is a generic function to extract user input values from a workflow or workflow run.
}
