% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{extract_inputs.workflow}
\alias{extract_inputs.workflow}
\title{Extract input values from a workflow}
\usage{
\method{extract_inputs}{workflow}(x, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A named list of input values extracted from the workflow steps.
}
\description{
This function extracts user input values from the steps of a workflow. It looks for parameters of type "input" in each step and collects their values into a named list.
}
