% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowrun-class.R
\name{extract_inputs.workflowrun}
\alias{extract_inputs.workflowrun}
\title{Extract input values from a workflow run}
\usage{
\method{extract_inputs}{workflowrun}(x, ...)
}
\arguments{
\item{x}{A `workflowrun` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A named list of input values.
}
\description{
This function extracts user input values used in the workflow run.
}
