% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-default-functions.R
\name{generate_letter_combinations}
\alias{generate_letter_combinations}
\title{Generate letter combinations}
\usage{
generate_letter_combinations(n_letters = 3, start = NULL, stop = NULL)
}
\arguments{
\item{n_letters}{An integer specifying the number of letters in each combination.
Must be a positive integer.}

\item{start}{A character string specifying the starting combination.
Must be of length `n_letters` and consist of lowercase letters.
If `NULL`, defaults to a string of 'a's (e.g., "aaa" for `n_letters = 3`).}

\item{stop}{A character string specifying the ending combination.
Must be of length `n_letters` and consist of lowercase letters.
If `NULL`, defaults to a string of 'z's (e.g., "zzz" for `n_letters = 3`).}
}
\value{
A character vector containing all combinations of letters from `start` to `stop`,
 inclusive, in lexicographical order.
}
\description{
This function generates all possible combinations of lowercase letters of a specified length,
starting from a given string and ending at another string. The combinations are generated
in lexicographical order. The function validates the input parameters to ensure they are
appropriate for generating the desired combinations.
}
