% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{get_field}
\alias{get_field}
\title{Get a specific field}
\usage{
get_field(x, field, ...)
}
\arguments{
\item{x}{The object to extract the field from.}

\item{field}{The name of the field to retrieve.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic get method for workflow-related objects.
}
