% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{get_field.workflow}
\alias{get_field.workflow}
\title{Get a specific field from a workflow}
\usage{
\method{get_field}{workflow}(x, field, step_name, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{field}{The name of the field to retrieve (e.g., "Name", "Comments").}

\item{step_name}{Name of the step to retrieve the field from.}

\item{...}{Additional arguments (not used).}
}
\value{
The value of the specified field from the workflow or a specific step.
}
\description{
Get a specific field from a workflow
}
