% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstep-class.R
\name{get_field.workflowstep}
\alias{get_field.workflowstep}
\title{Get a specific field from a workflowstep}
\usage{
\method{get_field}{workflowstep}(x, field, with_map_field = TRUE, ...)
}
\arguments{
\item{x}{A `workflowstep` object.}

\item{field}{The name of the field to retrieve (e.g., "name", "comments").}

\item{...}{Additional arguments (not used).}
}
\value{
The value of the specified field from the workflowstep.
}
\description{
Get a specific field from a workflowstep
}
