% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{goto_step.workflow}
\alias{goto_step.workflow}
\title{Go to a specific step in the workflow}
\usage{
\method{goto_step}{workflow}(x, index = NULL, id = NULL, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{index}{An integer index of the step to go to.}

\item{id}{An integer id of the step to go to.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflow` object with the current step set.
}
\description{
Go to a specific step in the workflow
}
