% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator
Returns left value if not NULL, otherwise right value.}
\usage{
a \%||\% b
}
\description{
Null coalescing operator
Returns left value if not NULL, otherwise right value.
}
