% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{import_workflow}
\alias{import_workflow}
\title{Import workflow from a ZIP file}
\usage{
import_workflow(zipfile, extract_dir)
}
\arguments{
\item{zipfile}{Path to the input ZIP file.}

\item{extract_dir}{Directory to extract the workflow files to.}
}
\value{
A `workflow` object created from the extracted files.
}
\description{
Import workflow from a ZIP file
}
