% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-inputs_table-module.R
\name{inputs_table_server}
\alias{inputs_table_server}
\title{Inputs Table Server}
\usage{
inputs_table_server(id, wf)
}
\arguments{
\item{id}{Module ID}

\item{wf}{Reactive expression containing the workflow object}
}
\value{
Shiny server logic for rendering the inputs table
}
\description{
Inputs Table Server
}
