% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-inputs_table-module.R
\name{inputs_table_ui}
\alias{inputs_table_ui}
\title{Inputs Table UI}
\usage{
inputs_table_ui(id, title = "")
}
\arguments{
\item{id}{Module ID}

\item{title}{Title to display above the table}
}
\value{
Shiny UI elements for the inputs table
}
\description{
Inputs Table UI
}
