% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstate-class.R
\name{new_workflowstate}
\alias{new_workflowstate}
\title{Create a new workflow state object}
\usage{
new_workflowstate(initial_input = NULL)
}
\arguments{
\item{initial_input}{The initial input provided to the workflow.}
}
\value{
A `workflowstate` object.
}
\description{
This object keeps track of the state while executing a workflow,
including the initial input, the last result, and the list of executed step runs.
}
