% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-utils.R
\name{null-coalescing-operator}
\alias{null-coalescing-operator}
\alias{\%||\%}
\title{Null coalescing operator
Returns left value if not NULL, otherwise right value.}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Left-hand side value.}

\item{b}{Right-hand side value.}
}
\value{
Left value if not NULL, otherwise right value.
}
\description{
Null coalescing operator
Returns left value if not NULL, otherwise right value.
}
