% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{previous_step.workflow}
\alias{previous_step.workflow}
\title{Move to the previous step in the workflow}
\usage{
\method{previous_step}{workflow}(x, wrap = FALSE, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{wrap}{Logical; if `TRUE`, wrap around to the last step when at the beginning.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflow` object with the current step moved back.
}
\description{
Move to the previous step in the workflow
}
