% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{print.workflow}
\alias{print.workflow}
\title{Print method for workflow objects}
\usage{
\method{print}{workflow}(x, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for workflow objects
}
