% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{print.workflowrun}
\alias{print.workflowrun}
\title{Print method for workflowrun objects}
\usage{
\method{print}{workflowrun}(x, ...)
}
\arguments{
\item{x}{A `workflowrun` object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for workflowrun objects
}
