% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstate-class.R
\name{print.workflowstate}
\alias{print.workflowstate}
\title{Print method for workflowstate objects}
\usage{
\method{print}{workflowstate}(x, ...)
}
\arguments{
\item{x}{A `workflowstate` object.}

\item{...}{Additional arguments (not used).}
}
\description{
Print method for workflowstate objects
}
