% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-default-functions.R
\name{prune_nulls}
\alias{prune_nulls}
\title{Recursively clean a list by removing NULL values
This function takes a list and recursively removes any NULL values from it.
If the input is not a list, it returns the input unchanged.
If the input is NULL, it returns NULL.}
\usage{
prune_nulls(x)
}
\arguments{
\item{x}{A list to be cleaned, or any other R object.}
}
\value{
A cleaned list with all NULL values removed, or the original input if it is not a list.
}
\description{
Recursively clean a list by removing NULL values
This function takes a list and recursively removes any NULL values from it.
If the input is not a list, it returns the input unchanged.
If the input is NULL, it returns NULL.
}
