% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{remove_step}
\alias{remove_step}
\title{Remove a step from a workflow}
\usage{
remove_step(workflow, step_name)
}
\arguments{
\item{workflow}{The workflow object from which the step will be removed.}

\item{step_name}{The name of the step to remove from the workflow.}
}
\value{
The updated workflow object with the specified step removed.
}
\description{
Generic method to remove a step from a workflow object.
}
