% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{remove_step.workflow}
\alias{remove_step.workflow}
\title{Remove a step from the workflow}
\usage{
\method{remove_step}{workflow}(x, position, ...)
}
\arguments{
\item{x}{The `workflow` object to update.}

\item{position}{An integer index specifying which step to remove.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflow` object with the specified step removed.
}
\description{
This method removes a `workflowstep` from the `steps` list of a `workflow` object at a specified
position. It also performs validation after removing the step and updates the current index and
entries of all steps to maintain consistency.
}
