% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-results_table-module.R
\name{results_table_ui}
\alias{results_table_ui}
\title{Results Table UI Module}
\usage{
results_table_ui(id)
}
\arguments{
\item{id}{Shiny module id}
}
\value{
Shiny UI output for the results table
}
\description{
Results Table UI Module
}
