% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{run}
\alias{run}
\title{Run a workflow step
This is a generic function to run a workflow step or an entire workflow.}
\usage{
run(object, state, ...)
}
\arguments{
\item{object}{A `workflowstep` object or a `workflow` object.}

\item{state}{A `workflowstate` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A `workflowsteprun` object representing the result of executing the step or
 a list containing the updated `workflow` and `workflowstate` objects.
}
\description{
Run a workflow step
This is a generic function to run a workflow step or an entire workflow.
}
