% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstep-class.R
\name{run.workflowstep}
\alias{run.workflowstep}
\title{Run a workflow step}
\usage{
\method{run}{workflowstep}(x, state, env = NULL, step_i = NULL, input_list = NULL, ...)
}
\arguments{
\item{x}{A `workflowstep` object representing the step to execute.}

\item{state}{A `workflowstate` object representing the current state of the workflow.}

\item{env}{An environment to look up the command function. Defaults to the step's
own env or the caller's env.}

\item{step_i}{The number of the step in the workflow, used for logging purposes.}

\item{input_list}{A list of inputs for argument parsing, loaded from the workflow's inputs file.}

\item{...}{Additional arguments (not used).}
}
\value{
A `workflowsteprun` object recording the execution of the step.
}
\description{
This function executes a single workflow step, updating the workflow state
with the result or error from the step execution.
}
