% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{run_workflow.workflow}
\alias{run_workflow.workflow}
\title{Run the entire workflow}
\usage{
\method{run_workflow}{workflow}(
  wf,
  state = list(),
  from = 1L,
  to = length(wf$steps),
  stop_on_error = TRUE,
  results_file = "results_summary.json",
  ...
)
}
\arguments{
\item{wf}{A `workflow` object.}

\item{state}{A `workflowstate` object representing the initial state.}

\item{from}{An integer index of the step to start from.}

\item{to}{An integer index of the step to end at.}

\item{stop_on_error}{Logical; if `TRUE`, stop execution on the first error.}

\item{results_file}{Name of the results summary file to update (default: "results_summary.json").}

\item{...}{Additional arguments passed to `run_step()`.}
}
\value{
A list containing the final workflow, state, and results of each step.
}
\description{
Run the entire workflow
}
