% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{save_as_zip}
\alias{save_as_zip}
\title{Save an object as a ZIP file}
\usage{
save_as_zip(x, file, ...)
}
\arguments{
\item{x}{The object to save.}

\item{file}{The path to the output ZIP file.}

\item{...}{Additional arguments (not used).}
}
\value{
None. The object is saved to the specified ZIP file.
}
\description{
This is a generic function to save various objects as ZIP files.
}
