% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{save_as_zip.workflow}
\alias{save_as_zip.workflow}
\title{Save workflow as a ZIP file}
\usage{
\method{save_as_zip}{workflow}(x, file, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{file}{Path to the output ZIP file.}

\item{...}{Additional arguments (not used).}
}
\description{
Save workflow as a ZIP file
}
