% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-functions.R
\name{simple_split}
\alias{simple_split}
\title{Split a single character by a pattern}
\usage{
simple_split(x, pattern, ...)
}
\arguments{
\item{x}{A character string to be split.}

\item{pattern}{A character string containing a regular expression to use for splitting.}

\item{...}{Additional arguments passed to `strsplit`.}
}
\value{
A character vector resulting from the split.
}
\description{
A simple wrapper around `strsplit` to split a single character string by a specified pattern.
}
