% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-startApplication.R
\name{startApplication}
\alias{startApplication}
\title{Start Application}
\usage{
startApplication(
  port = getOption("shiny.port"),
  host = "0.0.0.0",
  launch.browser = getOption("shiny.launch.browser", interactive())
)
}
\arguments{
\item{port}{port}

\item{host}{host}

\item{launch.browser}{If true, the system's default web browser will be launched}
}
\description{
Start Application
}
