% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstep-class.R
\name{summary.workflowsteprun}
\alias{summary.workflowsteprun}
\title{Summary method for workflowsteprun objects}
\usage{
\method{summary}{workflowsteprun}(object, ...)
}
\arguments{
\item{object}{A `workflowsteprun` object.}

\item{...}{Additional arguments (not used).}
}
\description{
Summary method for workflowsteprun objects
}
