% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{update}
\alias{update}
\title{Update an object}
\usage{
update(x, ...)
}
\arguments{
\item{x}{The object to update.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic update method for workflow-related objects.
}
