% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstate-class.R
\name{update.workflowstate}
\alias{update.workflowstate}
\title{Add a workflow step run to the workflow state}
\usage{
\method{update}{workflowstate}(x, steprun, idx, ...)
}
\arguments{
\item{x}{A `workflowstate` object.}

\item{steprun}{A `workflowsteprun` object to add.}

\item{idx}{The index at which to add the step run.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflowstate` object.
}
\description{
This function appends a `workflowsteprun` object to the list of executed step runs
}
