% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstep-class.R
\name{update.workflowstep}
\alias{update.workflowstep}
\title{Update a workflow step}
\usage{
\method{update}{workflowstep}(x, workflow_file_paths, value, field, with_map_field = TRUE, ...)
}
\arguments{
\item{x}{The `workflowstep` object to update.}

\item{workflow_file_paths}{The paths to the workflow files, used for updating related files.}

\item{value}{The new value to assign to the specified field.}

\item{field}{The name of the field to update. Must be one of "name", "label", "comments",
"command", "args", "iteration", "samples", or "loop".}

\item{with_map_field}{Logical, whether to map the field name using `map_field()`.
Defaults to `TRUE`.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflowstep` object.
}
\description{
This method allows updating specific fields of a `workflowstep` object, such as its name,
label, comments, command, parameters, or loop configuration.
}
