% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{update_functions}
\alias{update_functions}
\title{Update the custom functions of a workflow}
\usage{
update_functions(x, ...)
}
\arguments{
\item{x}{The object to update.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic method to update the custom functions file of a workflow object.
}
