% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-mutate.R
\name{update_functions.workflow}
\alias{update_functions.workflow}
\title{Update the functions file of a workflow}
\usage{
\method{update_functions}{workflow}(x, new_functions, ...)
}
\arguments{
\item{x}{The `workflow` object to update.}

\item{new_functions}{new functions}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflow` object with the new functions file content.
}
\description{
This method updates the content of the functions file associated with a
`workflow` object. If the workflow is file-backed, the updated content
will be written to the functions file.
}
