% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{update_input_list}
\alias{update_input_list}
\title{Update the input list of a workflow}
\usage{
update_input_list(x, ...)
}
\arguments{
\item{x}{The object to update.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Generic method to update the input list of a workflow object.
}
