% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-from-file.R
\name{workflow_example}
\alias{workflow_example}
\title{Workflow example from package files}
\usage{
workflow_example(name = "example_workflow")
}
\arguments{
\item{name}{Name for the workflow.}
}
\value{
A `workflow` object.
}
\description{
Workflow example from package files
}
