% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{workflow_file_paths}
\alias{workflow_file_paths}
\title{Helper to get file paths for PEITHO workflow files}
\usage{
workflow_file_paths(
  path = "",
  inputs = "",
  commands = "",
  results = "",
  functions = "",
  config_path = system.file("config", "config.json", package = "PEITHO")
)
}
\arguments{
\item{path}{Path to folder containing workflow files (default: PEITHO example folder).}

\item{inputs}{Name of the inputs file (default: "inputs.json").}

\item{commands}{Name of the commands file (default: "commands.json").}

\item{results}{Name of the results summary file (default: "results.json").}

\item{functions}{Name of the R script file containing custom functions (default: "functions.R").}

\item{config_path}{Path to configuration JSON file (default: package's config.json).}
}
\value{
A list with full paths to the specified files.
}
\description{
This function constructs full file paths for the workflow files located in a specified folder.
Validation of file existence is performed in `extract_workflow_from_files()`.
}
