% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow_files-module.R
\name{workflow_files_server}
\alias{workflow_files_server}
\title{Workflow Files Server}
\usage{
workflow_files_server(id, active_dir)
}
\arguments{
\item{id}{Module ID}

\item{active_dir}{Reactive expression containing the active workflow directory path}
}
\value{
Shiny server logic for rendering the workflow file tree
}
\description{
Workflow Files Server
}
