% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow_files-module.R
\name{workflow_files_ui}
\alias{workflow_files_ui}
\title{Workflow Files UI}
\usage{
workflow_files_ui(id, title = "")
}
\arguments{
\item{id}{Module ID}

\item{title}{Title to display above the file tree}
}
\value{
Shiny UI elements for the workflow files tab
}
\description{
Workflow Files UI
}
