% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow-graph-module.R
\name{workflow_graph_server}
\alias{workflow_graph_server}
\title{Workflow graph server module
This module renders a PEITHO workflow as a graph using DiagrammeR. It listens for changes to the workflow object and updates the graph accordingly.}
\usage{
workflow_graph_server(id, wf)
}
\arguments{
\item{id}{A unique identifier for the module namespace.}

\item{wf}{A reactive expression that returns the current workflow object to visualize.}
}
\value{
Server logic to render the workflow graph.
}
\description{
Workflow graph server module
This module renders a PEITHO workflow as a graph using DiagrammeR. It listens for changes to the workflow object and updates the graph accordingly.
}
