% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow-graph-module.R
\name{workflow_graph_ui}
\alias{workflow_graph_ui}
\title{Graph UI}
\usage{
workflow_graph_ui(id, title = "Workflow graph")
}
\arguments{
\item{id}{Module ID}

\item{title}{Title to display above the graph (default: "Workflow graph").}
}
\value{
Shiny UI elements for the workflow graph
}
\description{
Graph UI
}
