% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow_table-module.R
\name{workflow_table_server}
\alias{workflow_table_server}
\title{Workflow Table Server}
\usage{
workflow_table_server(id, wf)
}
\arguments{
\item{id}{Module ID}

\item{wf}{Reactive expression containing the workflow object}
}
\value{
Shiny server logic for rendering the workflow table
}
\description{
Workflow Table Server
}
