% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow_table-module.R
\name{workflow_table_ui}
\alias{workflow_table_ui}
\title{Workflow Table UI}
\usage{
workflow_table_ui(id, title = "")
}
\arguments{
\item{id}{Module ID}

\item{title}{Title to display above the table}
}
\value{
Shiny UI elements for the workflow table
}
\description{
Workflow Table UI
}
