% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-mutate.R
\name{write_functions_file}
\alias{write_functions_file}
\title{Write content to the functions file of a workflow}
\usage{
write_functions_file(dir_path, content)
}
\arguments{
\item{dir_path}{The directory path of the workflow. If NULL or non-existent,
no write will occur.}

\item{content}{The content to write to the functions file.}
}
\value{
NULL if the directory path is invalid; otherwise,
        the content is written to the functions file.
}
\description{
This helper function writes the provided content to the functions file
of a workflow, if the workflow is file-backed and has a valid functions path.
If the directory path is NULL or does not exist, the function will return
NULL without performing any write operation.
}
