% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{add_step}
\alias{add_step}
\title{Add a step to a workflow}
\usage{
add_step(workflow, step)
}
\arguments{
\item{workflow}{The workflow object to which the step will be added.}

\item{step}{The step to add to the workflow.}
}
\value{
The updated workflow object with the new step added.
}
\description{
Generic method to add a step to a workflow object.
}
