% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstate-class.R
\name{as.data.frame.workflowstate}
\alias{as.data.frame.workflowstate}
\title{Convert a workflowstate to a data frame}
\usage{
\method{as.data.frame}{workflowstate}(x, ...)
}
\arguments{
\item{x}{A `workflowstate` object.}

\item{...}{Additional arguments (not used).}
}
\value{
A data frame summarizing the workflow state.
}
\description{
This function summarizes the workflow state by converting the list of step runs
into a data frame.
}
