% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-operationparam-class.R
\name{extract_arg.operationparam}
\alias{extract_arg.operationparam}
\title{Extract argument value from operationparam object}
\usage{
\method{extract_arg}{operationparam}(x, result_path, ...)
}
\arguments{
\item{x}{A `operationparam` object.}

\item{result_path}{Path to the results JSON file.}

\item{...}{Additional arguments (not used).}
}
\value{
A list containing the argument value, named if applicable.
}
\description{
Extract argument value from operationparam object
}
