% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowrun-class.R
\name{extract_inputs}
\alias{extract_inputs}
\title{Extract input values from a workflow run}
\usage{
extract_inputs(wfr)
}
\arguments{
\item{wfr}{A `workflowrun` object.}
}
\value{
A named list of input values.
}
\description{
This function extracts user input values used in the workflow run.
}
