% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-from-file.R
\name{extract_workflow_from_files}
\alias{extract_workflow_from_files}
\title{Extract workflow steps from files in a folder}
\usage{
extract_workflow_from_files(
  path_to_folder = system.file("scripts", "peitho_files", package = "PEITHO")
)
}
\arguments{
\item{path_to_folder}{Path to folder containing `commands.json` and `inputs.txt` and `results_summary.json`.
Default is the package's `peitho_files` folder.}
}
\value{
A list of `workflowstep` objects.
}
\description{
Extract workflow steps from files in a folder
}
