% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{new_workflow}
\alias{new_workflow}
\title{Create a new workflow object}
\usage{
new_workflow(
  name = "Untitled workflow",
  steps = list(),
  current = if (length(steps)) 1L else NA_integer_,
  use_peitho_folder = TRUE,
  path_to_folder = system.file("scripts", "peitho_files", package = "PEITHO"),
  script = NULL,
  ...
)
}
\arguments{
\item{name}{A name for the workflow.}

\item{steps}{A list of `workflowstep` objects defining the steps of the workflow.}

\item{current}{The index of the current step in the workflow.}

\item{use_peitho_folder}{Logical; if `TRUE`, load steps from PEITHO example folder.}

\item{path_to_folder}{Path to folder containing `commands.json` and `inputs.txt` and `results_summary.json`.
Default is the package's `peitho_files` folder.}

\item{script}{An optional R script file defining custom functions for the workflow,
stored in the `dots` field.}

\item{...}{Additional metadata to store with the workflow.}
}
\value{
A `workflow` object.
}
\description{
This object represents a workflow, consisting of a sequence of workflow steps,
the current step index, and additional metadata.
}
