% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{new_workflow}
\alias{new_workflow}
\title{Create a new workflow object}
\usage{
new_workflow(
  steps = list(),
  name = "Untitled workflow",
  current = if (length(steps)) 1L else NA_integer_,
  ...
)
}
\arguments{
\item{steps}{A list of `workflow_step` objects defining the steps of the workflow.}

\item{name}{A name for the workflow.}

\item{current}{The index of the current step in the workflow.}

\item{...}{Additional metadata to store with the workflow.}
}
\value{
A `workflow` object.
}
\description{
This object represents a workflow, consisting of a sequence of workflow steps,
the current step index, and additional metadata.
}
