% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{new_workflow}
\alias{new_workflow}
\title{Create a new workflow object}
\usage{
new_workflow(
  name = "Untitled workflow",
  steps = list(),
  current = if (length(steps)) 1L else NA_integer_,
  use_peitho_folder = TRUE,
  workflow_file_paths = list(),
  ...
)
}
\arguments{
\item{name}{A name for the workflow.}

\item{steps}{A list of `workflowstep` objects defining the steps of the workflow.}

\item{current}{The index of the current step in the workflow.}

\item{use_peitho_folder}{Logical; if `TRUE`, load steps from PEITHO example folder.}

\item{workflow_file_paths}{A list of file paths for workflow files (see `workflow_file_paths()`).}

\item{...}{Additional metadata to store with the workflow.}
}
\value{
A `workflow` object.
}
\description{
This object represents a workflow, consisting of a sequence of workflow steps,
the current step index, and additional metadata.
}
