% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowrun-class.R
\name{new_workflowrun}
\alias{new_workflowrun}
\title{Workflow run object}
\usage{
new_workflowrun(workflow, state)
}
\arguments{
\item{workflow}{A `workflow` object.}

\item{state}{A `workflowstate` object.}
}
\value{
A `workflowrun` object.
}
\description{
This object represents a completed run of a workflow, containing the workflow definition
and the final state after execution.
}
