% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstep-class.R
\name{new_workflowstep}
\alias{new_workflowstep}
\title{Create a new workflow step object}
\usage{
new_workflowstep(
  entry,
  command,
  name = NULL,
  label = NULL,
  comments = "",
  args = "",
  params = list(),
  loop = "",
  env = parent.frame(),
  ...
)
}
\arguments{
\item{entry}{An integer representing the step's position in the workflow.
This should be unique for each step.}

\item{command}{A character string specifying the name of the function to execute for
this step, e.g. "strsplit". This function must exist in the loaded
name space or in a custom script environment.}

\item{name}{A human-readable name for the step. Defaults to "Step <entry>".}

\item{label}{A label for the step, used in UIs. Defaults to the same as `name`.}

\item{comments}{A character string with comments or description for the step.}

\item{args}{The original argument string from the workflow file, for reference.}

\item{params}{A list of parameters to pass to the command function.}

\item{loop}{A character string indicating if the step should be looped over.
Can be "yes", "no", or "auto".}

\item{env}{An environment to look up the command function. Default is the parent
frame.}

\item{...}{Additional metadata to store with the step.}
}
\value{
A `workflowstep` object.
}
\description{
This object defines a single step within a workflow,
including its unique identifier, name, command, parameters, and other metadata.
}
