% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflowstep-class.R
\name{new_workflowstep}
\alias{new_workflowstep}
\title{Create a new workflow step object}
\usage{
new_workflowstep(
  id,
  name = NULL,
  label = NULL,
  operation = NULL,
  params = list(),
  include_last = TRUE,
  source_external = "None",
  ...
)
}
\arguments{
\item{id}{An integer identifier for the step.}

\item{name}{A human-readable name for the step. Defaults to "Step <id>".}

\item{label}{A label for the step, used in UIs. Defaults to the same as `name`.}

\item{operation}{A character string specifying the name of the function to execute for
this step, e.g. "strsplit".
This function must exist in the loaded name space.}

\item{params}{A list of parameters to pass to the operation function.}

\item{include_last}{A logical indicating whether to include the last result
from the workflow state as an input to this step's operation.
Default is `TRUE`.}

\item{source_external}{A character string indicating the source of external data for this step.
Default is `"None"`.}

\item{...}{Additional metadata to store with the step.}
}
\value{
A `workflowstep` object.
}
\description{
This object defines a single step within a workflow,
including its unique identifier, name, operation, parameters, and other metadata.
}
