% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{remove_step}
\alias{remove_step}
\title{Remove a step from a workflow}
\usage{
remove_step(x, position, ...)
}
\arguments{
\item{x}{The workflow object from which the step will be removed.}

\item{position}{The position of the step to remove from the workflow.}

\item{...}{Additional arguments passed to methods.}
}
\value{
The updated workflow object with the specified step removed.
}
\description{
Generic method to remove a step from a workflow object.
}
