% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-results_table-module.R
\name{results_table_server}
\alias{results_table_server}
\title{Results Table Server Module}
\usage{
results_table_server(id, wf_run)
}
\arguments{
\item{id}{Shiny module id}

\item{wf_run}{reactive workflow run object}
}
\value{
None. Registers output$results_table
}
\description{
Results Table Server Module
}
