% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{run.workflow}
\alias{run.workflow}
\title{Run the entire workflow}
\usage{
\method{run}{workflow}(
  object,
  state = list(),
  from = 1L,
  to = length(object$steps),
  env = parent.frame(),
  ...
)
}
\arguments{
\item{object}{A `workflow` object.}

\item{state}{A `workflowstate` object representing the initial state.}

\item{from}{An integer index of the step to start from.}

\item{to}{An integer index of the step to end at.}

\item{env}{An environment to look up operation functions. Default is the parent frame.}

\item{...}{Additional arguments passed to `run.workflowstep()`.}
}
\value{
A list containing the final workflow, state, and results of each step.
}
\description{
Run the entire workflow
}
