% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{run.workflow}
\alias{run.workflow}
\title{Run the entire workflow}
\usage{
\method{run}{workflow}(x, state = list(), from = 1L, to = length(x$steps), env = NULL, ...)
}
\arguments{
\item{x}{A `workflow` object.}

\item{state}{A `workflowstate` object representing the initial state.}

\item{from}{An integer index of the step to start from.}

\item{to}{An integer index of the step to end at.}

\item{env}{An environment to look up operation functions. Defaults to `NULL`, which uses
each step's own env or the caller's env.}

\item{...}{Additional arguments passed to `run.workflowstep()`.}
}
\value{
A list containing the final workflow, state, and results of each step.
}
\description{
Run the entire workflow
}
