% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-startApplication.R
\name{startApplication}
\alias{startApplication}
\title{Start Application}
\usage{
startApplication(
  port = base::getOption("shiny.port"),
  launch.browser = base::getOption("shiny.launch.browser", base::interactive())
)
}
\arguments{
\item{port}{port}

\item{launch.browser}{If true, the system's default web browser will be launched}
}
\description{
Start Application
}
