% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{update}
\alias{update}
\title{Update a workflow state after running a step}
\usage{
update(x, steprun, idx, ...)
}
\arguments{
\item{x}{A `workflowstate` object.}

\item{steprun}{A `workflowsteprun` object representing the result of the executed step.}

\item{idx}{The index of the step run to update.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflowstate` object.
}
\description{
This is a generic function to update a workflow state after executing a step.
}
