% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{update.workflow}
\alias{update.workflow}
\title{Update a workflow step}
\usage{
\method{update}{workflow}(x, step, entry, value, ...)
}
\arguments{
\item{x}{A `workflowstep` object to update.}

\item{step}{The index of the step to update.}

\item{entry}{The name of the entry to update (one of "name", "label", "comments", "operation",
"params", "loop")}

\item{value}{The new value to assign to the specified entry.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflowstep` object.
}
\description{
This function updates a specific entry of a `workflowstep` object with a new value. It is used
to modify step details such as name, label, comments, operation, parameters, or loop settings.
}
