% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-workflow-class.R
\name{update_input_list.workflow}
\alias{update_input_list.workflow}
\title{Update the input list of a workflow}
\usage{
\method{update_input_list}{workflow}(x, new_list, write_file = TRUE, rebuild_params = TRUE, ...)
}
\arguments{
\item{x}{The `workflow` object to update.}

\item{new_list}{A named list of input values to update in the workflow.}

\item{write_file}{Logical; if `TRUE`, the updated input list will be written to the
inputs file if the workflow has associated file paths.}

\item{rebuild_params}{Logical; if `TRUE`, the parameters of the workflow steps will be
rebuilt based on the new input list using the existing argument strings and loop settings.}

\item{...}{Additional arguments (not used).}
}
\value{
The updated `workflow` object with the new input list and optionally updated parameters.
}
\description{
This method updates the `input_list` of a `workflow` object and optionally writes the
updated list to the corresponding inputs file if the workflow is file-backed. It also
has an option to rebuild the parameters of the workflow steps based on the new input list.
}
