% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow-graph-module.R
\name{workflow_graph_server}
\alias{workflow_graph_server}
\title{Graph Server}
\usage{
workflow_graph_server(id, wf)
}
\arguments{
\item{id}{Module ID}

\item{wf}{Reactive expression containing the workflow object}
}
\value{
Shiny server logic for rendering the workflow graph
}
\description{
Graph Server
}
