% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-workflow-graph-module.R
\name{workflow_graph_ui}
\alias{workflow_graph_ui}
\title{Workflow graph UI module
This module provides a UI component and server logic to visualize a PEITHO workflow as a graph using DiagrammeR.}
\usage{
workflow_graph_ui(id, title = "Workflow graph")
}
\arguments{
\item{id}{A unique identifier for the module namespace.}

\item{title}{The title to display above the graph (default: "Workflow graph").}

\item{wf}{A reactive expression that returns the current workflow object to visualize.}
}
\value{
A Shiny UI component for the workflow graph and server logic to render it.
}
\description{
Workflow graph UI module
This module provides a UI component and server logic to visualize a PEITHO workflow as a graph using DiagrammeR.
}
